+++ src/components/svgIcon/index.vue
@@ -1,15 +1,17 @@
+<template>
+	<div class="svg-icon-container">
+		<i v-if="isShowIconSvg" class="el-icon" :style="setIconSvgStyle">
+			<component :is="getIconName" />
+		</i>
+		<img v-else-if="isShowIconImg" :src="getIconName" :style="setIconImgStyle" />
+		<i v-else :class="getIconName" :style="setIconSvgStyle" />
+	</div>
+</template>
+
 <script lang="ts">
-// 渲染函数：https://v3.cn.vuejs.org/guide/render-function.html
-import { h, resolveComponent } from 'vue';
+import { computed, defineComponent } from 'vue';
 
-// 定义接口来定义对象的类型
-interface SvgIconProps {
-	name: string;
-	size: number;
-	color: string;
-}
-
-export default {
+export default defineComponent({
 	name: 'svgIcon',
 	props: {
 		// svg 图标组件名字
@@ -26,17 +28,43 @@ export default {
 			type: String,
 		},
 	},
-	setup(props: SvgIconProps) {
-		// 定义变量
-		const linesString: any[] = ['https', 'http', '/src', '/assets', import.meta.env.VITE_PUBLIC_PATH];
-		const onLineStyle: string = `font-size: ${props.size}px;color: ${props.color}`;
-		const localsStyle: string = `width: ${props.size}px;height: ${props.size}px`;
-		const eleSetStyle = { class: 'el-icon', style: onLineStyle };
+	setup(props) {
+		// 在线链接、本地引入地址前缀
+		const linesString = ['https', 'http', '/src', '/assets', import.meta.env.VITE_PUBLIC_PATH];
 
-		// 逻辑判断
-		if (props.name?.startsWith('ele-')) return () => h('i', eleSetStyle, [props.name === 'ele-' ? '' : h(resolveComponent(props.name))]);
-		else if (linesString.find((str) => props.name?.startsWith(str))) return () => h('img', { src: props.name, style: localsStyle });
-		else return () => h('i', { class: props.name, style: onLineStyle });
+		// 获取 icon 图标名称
+		const getIconName = computed(() => {
+			return props?.name;
+		});
+		// 用于判断 element plus 自带 svg 图标的显示、隐藏
+		const isShowIconSvg = computed(() => {
+			return props?.name?.startsWith('ele-');
+		});
+		// 用于判断在线链接、本地引入等图标显示、隐藏
+		const isShowIconImg = computed(() => {
+			return linesString.find((str) => props.name?.startsWith(str));
+		});
+		// 设置图标样式
+		const setIconSvgStyle = computed(() => {
+			return `font-size: ${props.size}px;color: ${props.color};height: 100%;display: flex;align-items: center;`;
+		});
+		// 设置图片样式
+		const setIconImgStyle = computed(() => {
+			return `width: ${props.size}px;height: ${props.size}px`;
+		});
+		return {
+			getIconName,
+			isShowIconSvg,
+			isShowIconImg,
+			setIconSvgStyle,
+			setIconImgStyle,
+		};
 	},
-};
+});
 </script>
+
+<style scoped lang="scss">
+.svg-icon-container {
+	display: inline-block;
+}
+</style>